+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Export / import documents, documents types, datatypes, views, layout and scripts ([https://github.com/GotCms/GotCms/issues/146](https://github.com/GotCms/GotCms/issues/146))
 - Display property description with datatypes ([https://github.com/GotCms/GotCms/issues/156](https://github.com/GotCms/GotCms/issues/156))
 - Update bootstrap and lessjs ([https://github.com/GotCms/GotCms/issues/157](https://github.com/GotCms/GotCms/issues/157))
+- Bug with arcana template ([https://github.com/GotCms/GotCms/issues/158](https://github.com/GotCms/GotCms/issues/158))
 
 ##1.0.0 (28 Oct 2013):
 - Bug with french translation ([https://github.com/GotCms/GotCms/issues/131](https://github.com/GotCms/GotCms/issues/131))